Moved footer out of the index page
