+++ components/footer/footer.style.js
@@ -1,0 +1,4 @@
+import { makeStyles } from '@material-ui/core/styles'
+
+export const useStyles = makeStyles(() => ({
+}))
