+++ components/index.js
@@ -1,3 +1,4 @@
 export * from './search-box/search-box'
 export * from './options-box/options-box'
 export * from './result-table/result-table'
+export * from './footer/footer'
